Move validate back to validators
